src/agent/docker-rest-agent/requirements.txt
Django>=3.0
uwsgi
enum34
djangorestframework
holdup>1.5.0,<=1.6.0
drf-yasg<=1.17.0
swagger_spec_validator<=2.4.1
psycopg2-binary
celery<5.0,>=4.4
redis
requests
supervisor
django-celery-beat
django-celery-results
django-3-jet
djangorestframework-jwt<=1.11.0
python-jwt # 需要安装，否则会出现token解码失败错误
shortuuid
